requirements.txt
streamlit==1.28.0
paddlepaddle==2.6.2
paddleocr==2.7.0.0
opencv-python-headless==4.6.0.66
numpy==1.24.3
Pillow==10.0.1


